Implemented Activity Command decorator as abstract entity
