+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/di/modules/NavigationModule.java
@@ -1,14 +1,11 @@
 package sasd97.java_blog.xyz.yandexweather.di.modules;
 
-import android.content.Context;
-import android.util.SparseArray;
-
 import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
 import sasd97.java_blog.xyz.yandexweather.navigation.activities.ActivityCommand;
-import sasd97.java_blog.xyz.yandexweather.navigation.activities.AppActivityRouter;
+import sasd97.java_blog.xyz.yandexweather.navigation.AppActivityRouter;
 import sasd97.java_blog.xyz.yandexweather.navigation.Router;
 
 /**
