+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/navigation/activities/MoveForward.java
@@ -9,33 +9,17 @@ import java.lang.ref.WeakReference;
  * Created by alexander on 09/07/2017.
  */
 
-public class MoveForward implements ActivityCommandDecorator {
+public class MoveForward extends ActivityCommandDecorator {
 
-    private WeakReference<Activity> currentActivity;
     private Class<?> destinationActivity;
 
-    private ActivityCommand wrapped;
-
     public MoveForward(Activity currentActivity, Class<?> destinationActivity) {
-        this.currentActivity = new WeakReference<>(currentActivity);
+        super(currentActivity);
         this.destinationActivity = destinationActivity;
     }
 
     @Override
-    public void setNext(ActivityCommand wrapped) {
-        this.wrapped = wrapped;
-    }
-
-    @Override
-    public void apply() {
-        Activity activity = currentActivity.get();
-        if (activity == null) return;
+    protected void onApply(Activity activity) {
         activity.startActivity(obtainIntent(activity, destinationActivity));
-
-        if (wrapped != null) wrapped.apply();
-    }
-
-    private Intent obtainIntent(Activity from, Class<?> to) {
-        return new Intent(from, to);
     }
 }
