+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/navigation/fragments/FragmentCommandDecorator.java
@@ -38,5 +38,5 @@ public abstract class FragmentCommandDecorator implements FragmentCommand {
         return t;
     }
 
-    public abstract FragmentTransaction onApply(FragmentTransaction transaction);
+    protected abstract FragmentTransaction onApply(FragmentTransaction transaction);
 }
