+++ app/src/main/java/sasd97/java_blog/xyz/yandexweather/navigation/fragments/Replace.java
@@ -18,7 +18,7 @@ public class Replace extends FragmentCommandDecorator {
     }
 
     @Override
-    public FragmentTransaction onApply(FragmentTransaction transaction) {
+    protected FragmentTransaction onApply(FragmentTransaction transaction) {
         return transaction.replace(getContainer(), destinationFragment);
     }
 }
